VB_Schedule_Nebraska/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
playwright==1.47.2
pdfplumber==0.11.4
